+++ packages/core/src/constants/z-index.js
@@ -1,0 +1,13 @@
+export default {
+  loadingCover: 6000,
+  popup: 5000,
+  hamburgerMenu: 4000,
+  toc: 3000, // table of contents
+  webPush: 2000,
+  buttomBanner: 2000,
+  header: 1000,
+  articleFullScreenHeroImage: 900,
+  embedUp: 800,
+  mobileToolBar: 700,
+  embedDown: 600,
+}